Added python code for the chatbot
code defines the functions we need to construct the index and query it
You need an OPENAI API key to be able to run this code.
If you don't have one yet, get it by signing up. Then click your account icon on the top right of the screen and select "View API Keys". Create an API key.
This will also take every file in the folder 'data', split it into chunks, and embed it with OpenAI's embeddings API.
Notice: running this code will cost you credits on your OpenAPI account ($0.02 for every 1,000 tokens). If you've just set up your account, the free credits that you have should be more than enough for this experiment.